added search dialog
